more work

+++ crates/cabin-language-server/src/lsp/text_document/mod.rs
@@ -48,7 +48,7 @@ impl Position {
 		let mut column = 0;
 		for (position, character) in text.chars().enumerate() {
 			if line as usize == self.line && column as usize == self.character {
-				return Span::new(position as usize, 1);
+				return Span::new(position, 1);
 			}
 
 			column += 1;
